mcp-server/requirements.txt
mcp>=1.3.0
openai-whisper>=20231117
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
